Add javadoc for frontend divided callbacks botton into packages and divided the swagger docs into packages
