+++ src/main/java/com/evgeniyfedorchenko/animalshelter/admin/annotations/documentation/adopter/SearchAdoptersDocumentation.java
@@ -1,4 +1,4 @@
-package com.evgeniyfedorchenko.animalshelter.admin.annotations.documentation;
+package com.evgeniyfedorchenko.animalshelter.admin.annotations.documentation.adopter;
 
 import com.evgeniyfedorchenko.animalshelter.backend.dto.AdopterOutputDto;
 import io.swagger.v3.oas.annotations.Operation;
