+++ src/main/java/com/evgeniyfedorchenko/animalshelter/admin/annotations/documentation/animal/AssignAnimalToAdopterDocumentation.java
@@ -1,4 +1,4 @@
-package com.evgeniyfedorchenko.animalshelter.admin.annotations.documentation;
+package com.evgeniyfedorchenko.animalshelter.admin.annotations.documentation.animal;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.links.Link;
