+++ src/main/java/com/evgeniyfedorchenko/animalshelter/backend/services/TelegramService.java
@@ -1,20 +1,8 @@
 package com.evgeniyfedorchenko.animalshelter.backend.services;
 
-import com.evgeniyfedorchenko.animalshelter.telegram.listener.TelegramBot;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
-@Service
-public class TelegramService {
+public interface TelegramService {
 
-    @Autowired
-    private TelegramBot telegramBot;
-
-    public boolean sendMessage(String chatId, String message) throws TelegramApiException {
-        SendMessage sendMessage = new SendMessage(chatId, message);
-        telegramBot.execute(sendMessage);
-        return true;
-    }
+    boolean sendMessage(String chatId, String message) throws TelegramApiException;
 }
