+++ src/main/java/com/evgeniyfedorchenko/animalshelter/backend/services/TelegramServiceImpl.java
@@ -1,0 +1,21 @@
+package com.evgeniyfedorchenko.animalshelter.backend.services;
+
+import com.evgeniyfedorchenko.animalshelter.telegram.listener.TelegramBot;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+@Service
+public class TelegramServiceImpl implements TelegramService {
+
+    @Autowired
+    private TelegramBot telegramBot;
+
+    @Override
+    public boolean sendMessage(String chatId, String message) throws TelegramApiException {
+        SendMessage sendMessage = new SendMessage(chatId, message);
+        telegramBot.execute(sendMessage);
+        return true;
+    }
+}
