+++ src/main/java/com/evgeniyfedorchenko/animalshelter/telegram/configuration/BotConfiguration.java
@@ -8,10 +8,19 @@ import org.telegram.telegrambots.meta.TelegramBotsApi;
 import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 import org.telegram.telegrambots.updatesreceivers.DefaultBotSession;
 
+/**
+ * Telegram bot configuration class
+ */
 @Slf4j
 @Configuration
 public class BotConfiguration {
 
+    /**
+     * A method that configures the bin for <b>Spring context</b> based
+     * on the collected object {@link TelegramBot} and registers it on telegram servers
+     * @param telegramBot Initialized object to register
+     * @return System object for work on Telegram rules
+     */
     @Bean
     public TelegramBotsApi telegramBotsApi(TelegramBot telegramBot) {
