--- src/main/java/com/evgeniyfedorchenko/animalshelter/telegram/handler/buttons/Command.java
@@ -1,25 +1,0 @@
-package com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons;
-
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
-
-/**
- * Общий интерфейс для команд (и потенциально для обычных кнопок).
- * Команды, расширяющие этот интерфейс заканчиваются на -Command
- */
-public interface Command {
-
-    /**
-     * @return Возвращает строковое представление вызова. Так, как оно приходит из телеграма,
-     * чтобы можно было отловить по этому параметру
-     */
-    default String getTitle() {
-        return "/" + this.getClass().getSimpleName().toLowerCase();
-    }
-
-    /**
-     * Метод, который выполняет логику, заложенную в команду и формирует ответное сообщение
-     * @param chatId Идентификатор чата, в который будет отправлено сообщение
-     * @return готовый объект {@link SendMessage} для отправки
-     */
-    SendMessage apply(String chatId);
-}
