+++ src/main/java/com/evgeniyfedorchenko/animalshelter/telegram/handler/buttons/callbacks/GetPatternReport.java
@@ -1,0 +1,29 @@
+package com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.callbacks;
+
+import com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageModel;
+import com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageUtils;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageText;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+import static com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageData.*;
+
+@Component("GetPatternReport")
+public class GetPatternReport implements Callback {
+
+    @Override
+    public EditMessageText apply(String chatId, Integer messageId) {
+
+        Map<String, String> keyboardData = new LinkedHashMap<>();
+
+        keyboardData.put("Начнем! (backend)", START.getCallbackData());
+        keyboardData.put("Назад",             MAIN_REPORT_MENU.getCallbackData());
+
+        MessageUtils messageUtils = new MessageUtils();
+        MessageModel messageModel = new MessageModel(chatId, messageId, GET_PATTERN_REPORT, keyboardData);
+
+        return messageUtils.applyCallback(messageModel);
+    }
+}
