+++ src/main/java/com/evgeniyfedorchenko/animalshelter/telegram/handler/buttons/callbacks/SecurityContacts.java
@@ -1,0 +1,28 @@
+package com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.callbacks;
+
+import com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageModel;
+import com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageUtils;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.meta.api.methods.updatingmessages.EditMessageText;
+
+import java.util.LinkedHashMap;
+import java.util.Map;
+
+import static com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageData.MAIN_ABOUT_SHELTER;
+import static com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageData.SECURITY_CONTACTS;
+
+@Component("SecurityContacts")
+public class SecurityContacts implements Callback {
+
+    @Override
+    public EditMessageText apply(String chatId, Integer messageId) {
+
+        Map<String, String> keyboardData = new LinkedHashMap<>();
+        keyboardData.put("Назад", MAIN_ABOUT_SHELTER.getCallbackData());
+
+        MessageUtils messageUtils = new MessageUtils();
+        MessageModel messageModel = new MessageModel(chatId, messageId, SECURITY_CONTACTS, keyboardData);
+
+        return messageUtils.applyCallback(messageModel);
+    }
+}
