+++ src/main/java/com/evgeniyfedorchenko/animalshelter/telegram/handler/buttons/commands/Volunteer.java
@@ -1,0 +1,29 @@
+package com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.commands;
+
+import com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageModel;
+import com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageUtils;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Component;
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+
+import static com.evgeniyfedorchenko.animalshelter.telegram.handler.buttons.MessageData.VOLUNTEER;
+
+@Slf4j
+@Component("/volunteer")
+public class Volunteer implements Command {
+
+    @Override
+    public SendMessage apply(String chatId) {
+
+        new Thread(this::callVolunteer).start();
+
+        MessageUtils messageUtils = new MessageUtils();
+        MessageModel messageModel = new MessageModel(chatId, VOLUNTEER, null);
+        return messageUtils.applyCommand(messageModel);
+    }
+
+    private void callVolunteer() {
+        // todo Реализовать вызов волонтера
+        log.info("Calling volunteer...");
+    }
+}
