+++ src/main/java/com/evgeniyfedorchenko/animalshelter/telegram/listener/TelegramBot.java
@@ -1,15 +1,21 @@
 package com.evgeniyfedorchenko.animalshelter.telegram.listener;
 
 import com.evgeniyfedorchenko.animalshelter.telegram.handler.UpdateDistributor;
+import jakarta.validation.constraints.NotNull;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.objects.Update;
+import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
 import java.io.Serializable;
 
+/**
+ * A class representing a bot object registered and configured
+ * with a private token.Allows you to interact with Telegram servers
+ */
 @Slf4j
 @Component
 public class TelegramBot extends TelegramLongPollingBot {
@@ -27,6 +33,9 @@ public class TelegramBot extends TelegramLongPollingBot {
         return "animal_shelter_helper_bot";
     }
 
+    /**
+     * A method for receiving messages directly from the Telegram servers
+     */
     @Override
     public void onUpdateReceived(Update update) {
 
@@ -41,14 +50,19 @@ public class TelegramBot extends TelegramLongPollingBot {
         }
     }
 
-//    Нужно ли тут synchronized? Ну типа чтоб не перемешивались отправляемые сообщения
-//    из разных потоков, когда несколько людей обратились к боту одновременно
-    public synchronized void send(BotApiMethod<? extends Serializable> messToSend) {
+
+    /**
+     * A method for sending messages directly to the Telegram servers
+     * In case of an exception, it will be logged as {@code TelegramApiException was thrown. Cause: ex.getMessage()}
+     *
+     * @param messToSend @NotNull The object of the message ready to be sent
+     */
+    public void send(@NotNull BotApiMethod<? extends Serializable> messToSend) {
 
         try {
             execute(messToSend);
-        } catch (Exception ex) {
-            log.error(ex.getMessage());
+        } catch (TelegramApiException ex) {
+            log.error("TelegramApiException was thrown. Cause: {}", ex.getMessage());
         }
 
     }
